+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # SEOmatic Changelog
 
+## 3.1.35 - 2018.12.03
+### Changed
+* Fixed an issue where the `potentialAction` JSON-LD for the Site Links Search was rendered even if it was left empty
+* Fixed an issue where the SEO Settings field did not properly override the `mainEntityOfPage`
+
 ## 3.1.34 - 2018.11.28
 ### Added
 * Added a level 2 cache on the controller-based API requests for meta containers to improve "headless" performance